Update multithreading example.hta
In this update, in addition to bug fixes, we added an option to preserve image transparency in the Settings section of the Mix Images tab. Checking this option will cause transparent portions of both input images to remain transparent in the output composite images. Thanks to LesFerch for helping us on Stack Overflow to implement this feature.